add custom database configuration repository
