+++ config/settings.php
@@ -1,6 +1,4 @@
 <?php
-use App\Settings\GeneralSettings;
-use App\Settings\AccountSettings;
 
 return [
 
@@ -37,28 +35,10 @@ return [
      */
     'repositories' => [
         'database' => [
-            'type' => get_class(new class([
-                'model' => null,
-                'table' => null,
-                'connection' => null,
-            ]) extends \Spatie\LaravelSettings\SettingsRepositories\DatabaseSettingsRepository {
-                public function getPropertiesInGroup(string $group): array
-                {
-                    if (in_array(
-                            app(\Symfony\Component\Console\Input\ArgvInput::class)->getFirstArgument(), 
-                            ['migrate', 'package:discover', 'filament:upgrade', 'key:generate']
-                        )
-                    ) {
-                        return [];
-                    }
-
-                    return parent::getPropertiesInGroup($group);
-                }
-            }),
-            // 'type' => Spatie\LaravelSettings\SettingsRepositories\DatabaseSettingsRepository::class,
-            // 'model' => null,
-            // 'table' => null,
-            // 'connection' => null,
+            'type' => \App\Support\LaravelSettings\DatabaseSettingsRepository::class,
+            'model' => null,
+            'table' => null,
+            'connection' => null,
         ],
         'redis' => [
             'type' => Spatie\LaravelSettings\SettingsRepositories\RedisSettingsRepository::class,
